refactor: Update checkBalance import paths to use new balanceMethods module
