+++ api/app/clients/callbacks/createStartHandler.js
@@ -2,7 +2,7 @@ const { promptTokensEstimate } = require('openai-chat-tokens');
 const { EModelEndpoint, supportsBalanceCheck } = require('librechat-data-provider');
 const { formatFromLangChain } = require('~/app/clients/prompts');
 const { getBalanceConfig } = require('~/server/services/Config');
-const checkBalance = require('~/models/checkBalance');
+const checkBalance = require('models/balanceMethods');
 const { logger } = require('~/config');
 
 const createStartHandler = ({
