+++ api/server/controllers/assistants/chatV2.js
@@ -25,7 +25,7 @@ const { addTitle } = require('~/server/services/Endpoints/assistants');
 const { sendMessage, sleep, countTokens } = require('~/server/utils');
 const { createRunBody } = require('~/server/services/createRunBody');
 const { getTransactions } = require('~/models/Transaction');
-const checkBalance = require('~/models/checkBalance');
+const checkBalance = require('models/balanceMethods');
 const { getConvo } = require('~/models/Conversation');
 const getLogStores = require('~/cache/getLogStores');
 const { getModelMaxTokens } = require('~/utils');
